requirements.txt
pyaudio
pynput
openai
python-libxdo

